fix(router): satisfy type-discipline and cover session.model rewrite
Drop the redundant comment, mark the session copy as Final/mutable-ok,
and move the regression test into tests/test_litellm so CI coverage hits it.
