+++ tests/router_unit_tests/test_router_endpoints.py
@@ -521,42 +521,6 @@ def test_generic_api_call_with_fallbacks_basic(sync_mode):
 
 
 @pytest.mark.asyncio
-async def test_ageneric_api_call_with_fallbacks_resolves_session_model():
-    """
-    Regression test for https://github.com/BerriAI/litellm/issues/36742: when a
-    call has a nested session.model (realtime client_secrets), it must be
-    rewritten to the deployment litellm actually picked for the model_group,
-    not left as the caller's original alias.
-    """
-    mock_function = AsyncMock()
-    mock_function.__name__ = "acreate_realtime_client_secret"
-    mock_function.return_value = {"value": "ek_test"}
-
-    router = Router(
-        model_list=[
-            {
-                "model_name": "gpt-realtime-2-1-mini",
-                "litellm_params": {
-                    "model": "azure/gpt-realtime-2-1-mini-deployment",
-                    "api_key": "fake-api-key",
-                    "api_base": "https://fake.openai.azure.com",
-                },
-            }
-        ]
-    )
-
-    await router._ageneric_api_call_with_fallbacks(
-        model="gpt-realtime-2-1-mini",
-        original_function=mock_function,
-        session={"type": "realtime", "model": "gpt-realtime-2-1-mini"},
-    )
-
-    call_kwargs = mock_function.call_args.kwargs
-    assert call_kwargs["model"] == "azure/gpt-realtime-2-1-mini-deployment"
-    assert call_kwargs["session"]["model"] == "azure/gpt-realtime-2-1-mini-deployment"
-
-
-@pytest.mark.asyncio
 async def test_aadapter_completion():
     """
     Test the aadapter_completion method which uses async_function_with_fallbacks
